code corrections after merging (#16)
